requirements.txt
fastapi
uvicorn
pillow
pydantic
